Better checking for pointer dereferences
