+++ src/DwarfInstructions.hpp
@@ -15,10 +15,6 @@
 #include <stdint.h>
 #include <stdio.h>
 #include <stdlib.h>
-#include <errno.h>
-
-#include <unistd.h>
-#include <sys/syscall.h>
 
 #include "dwarf2.h"
 #include "Registers.hpp"
@@ -148,21 +144,6 @@ v128 DwarfInstructions<A, R>::getSavedVectorRegister(
     break;
   }
   _LIBUNWIND_ABORT("unsupported restore location for vector register");
-}
-
-
-// Check validity of the CFA.
-// This is a very dirty hack (inspired by original libunwind version).
-// Motivation: sometimes libunwind parse wrong value instead of CFA.
-// We check that memory address belongs to our process by issuing "mincore" syscall.
-// Actually we don't care if the address is in core or not, we only check return code.
-// If Address Sanitizer will argue, replace syscall to inline assembly.
-template <typename pint_t>
-static bool isPointerValid(pint_t ptr)
-{
-  unsigned char mincore_res = 0;
-  auto page_size = sysconf(_SC_PAGESIZE);
-  return ptr && (0 == syscall(SYS_mincore, (void*)(ptr / page_size * page_size), 1, &mincore_res) || errno == ENOSYS);
 }
 
 
@@ -320,12 +301,6 @@ DwarfInstructions<A, R>::evaluateExpression(pint_t expression, A &addressSpace,
     case DW_OP_deref:
       // pop stack, dereference, push result
       value = *sp--;
-      // Some libraries may have wrong DWARF expression (that's used to calculate CFA).
-      // Due to: - bug in compiler; - bug in manually written assembly code.
-      // Using this expression to dereference a pointer may cause segfault.
-      // Note: zero return value will be subsequently checked in the 'stepWithDwarf' function.
-      if (!isPointerValid(value))
-          return 0;
       *(++sp) = addressSpace.getP(value);
       if (log)
         fprintf(stderr, "dereference 0x%" PRIx64 "\n", (uint64_t)value);
@@ -790,8 +765,6 @@ DwarfInstructions<A, R>::evaluateExpression(pint_t expression, A &addressSpace,
     case DW_OP_deref_size:
       // pop stack, dereference, push result
       value = *sp--;
-      if (!isPointerValid(value))
-          return 0;
       switch (addressSpace.get8(p++)) {
       case 1:
         value = addressSpace.get8(value);
